Added method to get demo products (include products with lower sustainability scores).
