+++ app.py
@@ -5,7 +5,7 @@ from flask import Flask, request, jsonify
 from flask_cors import CORS
 
 from image_request.get_image import get_image_url_api
-from score_request.get_popular_products import get_popular_products
+from score_request.get_products import get_popular_products, get_demo_products
 from score_request.get_sustainability_score import get_sustainability_score
 from story_request.get_story import get_story_api
 from utils.check_env_variables import check_env_variables
@@ -75,3 +75,8 @@ def get_image_dummy():
 @app.route('/get_popular_products')
 def get_popular_prods():
     return get_popular_products(n_products=50)
+
+
+@app.route('/get_demo_products')
+def get_demo_prods():
+    return get_demo_products(n_products=50)
